requirements.txt
# Enterprise RAG System - Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# ML/AI Core
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
datasets==2.14.6
accelerate==0.24.1
peft==0.6.2

# Vector Databases
chromadb==0.4.18
pinecone-client==2.2.4
faiss-cpu==1.7.4

# LangChain
langchain==0.0.335
langchain-community==0.0.10

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
markdown==3.5.1

# Data & Analytics
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Database
sqlalchemy==2.0.23
sqlite3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tqdm==4.66.1
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Optional Cloud
boto3==1.34.0
google-cloud-storage==2.10.0
